fix(scout-for-lol): enforce the hooks-to-components direction in the web app
Why: review finding on #2423. The app's declared composition order is
`routes -> components -> hooks -> lib`, but the hooks rule only forbade
`routes/`. A hook could import UI and the check stayed green, so the gate did
not enforce the order the config claimed. It was already happening:
`hooks/use-explore-turn-actions.ts` imported `ExploreRunsContextValue` from
`components/explore-runs-context.ts`.

What:
- `StartExploreTurnInput` and `ExploreRunsContextValue` move to
  `lib/explore-runs-contract.ts`. They are contracts describing what an
  explore-run consumer can do, not React wiring; the hooks that act on a turn
  need the shape, not the context object. `components/explore-runs-context.ts`
  keeps the `createContext` call and the `useExploreRuns` accessor, and takes
  the type from `lib/`.
- `hooks-do-not-depend-on-routes` becomes
  `hooks-do-not-depend-on-routes-or-components`. Components call hooks, so a
  hook importing a component inverts the order and binds the hook to one
  rendering of the thing it manages.

Verification:
- `bunx turbo run typecheck lint test --filter=@scout-for-lol/app`: 213 tests
  pass; lint 0 errors; `check-architecture` reports 198 modules clean.
- Negative proof: re-adding
  `import "#src/components/explore-runs-context.ts"` to
  `hooks/use-explore-turn-actions.ts` now fails with
  `hooks-do-not-depend-on-routes-or-components`, where before the change the
  same edge passed.
