+++ packages/scout-for-lol/packages/app/architecture-fixtures/hooks-imports-routes.ts
@@ -1,4 +1,4 @@
-// Deliberate violation of hooks-do-not-depend-on-routes.
+// Deliberate violation of hooks-do-not-depend-on-routes-or-components.
 import "#src/routes/competition-list.tsx";
 
 export const illegalHooksDependency = true;
