+++ packages/scout-for-lol/packages/app/src/architecture-boundaries.test.ts
@@ -1,22 +1,10 @@
-import { describe, expect, it } from "vitest";
-import {
-  cruiseArchitectureFixtures,
-  expectedFixtureRuleNames,
-} from "@shepherdjerred/architecture";
+import { assertArchitectureFixtures } from "@shepherdjerred/architecture";
+import { test } from "vitest";
 import architecture from "#architecture";
 
-const packageRoot = import.meta.dir.replace(/\/src$/u, "");
-
-describe("dependency-cruiser layer boundaries", () => {
-  it("rejects a committed negative fixture for every declared boundary", async () => {
-    const result = await cruiseArchitectureFixtures({
-      packageRoot,
-      definition: architecture,
-    });
-
-    expect(result.violatedRuleNames).toEqual(
-      expectedFixtureRuleNames(architecture),
-    );
-    expect(result.errorCount).toBe(result.fixtureFiles.length);
+test("architecture fixtures prove every app boundary", async () => {
+  await assertArchitectureFixtures({
+    packageRoot: import.meta.dir.replace(/\/src$/u, ""),
+    definition: architecture,
   });
 });
