+++ packages/scout-for-lol/packages/app/src/hooks/use-explore-turn-actions.ts
@@ -1,7 +1,7 @@
 import { useCallback, type RefObject } from "react";
 import type { NavigateFunction } from "react-router";
 import type { ExploreMessage } from "@scout-for-lol/data";
-import type { ExploreRunsContextValue } from "#src/components/explore-runs-context.ts";
+import type { ExploreRunsContextValue } from "#src/lib/explore-runs-contract.ts";
 import { track } from "#src/lib/analytics.ts";
 import { shouldOpenStartedExploreConversation } from "#src/lib/explore-navigation.ts";
